made new widget for customeTabBar
