+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:tabbar/tabbar.dart';
+import 'custometabbar.dart';
 
 void main() => runApp(MyApp());
 
@@ -12,45 +12,8 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(),
+      home: CustomeTabBar(),
     );
   }
 }
 
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
-
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  final controller = PageController();
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Flutter Demo Home Page"),
-        centerTitle: true,
-        bottom: PreferredSize(
-          preferredSize: Size.fromHeight(kToolbarHeight),
-          child: TabbarHeader(
-            controller: controller,
-            tabs: [
-              Tab(icon: Icon(Icons.map), text: "Map"),
-              Tab(icon: Icon(Icons.list), text: "List"),
-              Tab(icon: Icon(Icons.person), text: "User profile", ),
-            ],
-          ),
-        ),
-      ),
-      body: TabbarContent(
-        controller: controller,
-        children: <Widget>[
-          Container(),
-          Container(),
-          Container(),
-        ],
-      ),
-    );
-  }
-}
